Added VerificationUpdater tests

+++ src/test/ssc_flairbot/VerificationUpdaterTest.java
@@ -1,22 +1,25 @@
 package ssc_flairbot;
 
+import no.stelar7.api.l4j8.basic.constants.api.Platform;
+import no.stelar7.api.l4j8.impl.L4J8;
+import no.stelar7.api.l4j8.impl.builders.spectator.SpectatorBuilder;
+import no.stelar7.api.l4j8.impl.builders.summoner.SummonerBuilder;
+import no.stelar7.api.l4j8.pojo.spectator.SpectatorParticipant;
+import no.stelar7.api.l4j8.pojo.summoner.Summoner;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import ssc_flairbot.league.LeagueApi;
 import ssc_flairbot.persistence.DBHandler;
 import ssc_flairbot.persistence.User;
 import ssc_flairbot.persistence.UserBuilder;
-import ssc_flairbot.reddit.FlairHandler;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.junit.Assert.*;
 
@@ -29,39 +32,68 @@ public class VerificationUpdaterTest {
     @Autowired
     VerificationUpdater verificationUpdater;
     @Autowired
-    LeagueApi lolApi;
-    @Autowired
     DBHandler database;
+
+    private static boolean setUpIsDone = false;
+    private final L4J8 riotApi = new L4J8(SecretFile.CREDS);
+    private static List<User> users = new ArrayList<>();
 
     @Before
     public void setUp() {
-        database.dropTable();
-        database.createTable();
-
-        User user1 = new UserBuilder().redditName("Thorasine").summonerName("Trefort")
-                .summonerId("8kFIUtL2QHnAKmyI485jY7bWifUk6poPC1KQehEbjtr6zCc").server("EUW").validated("validated")
-                .validationCode("83ERFK").validationTries(0).buildUser();
-
-        User user2 = new UserBuilder().redditName("Thorasine").summonerName("Oreena")
-                .summonerId("YKuxOWZtcpkiE1V07iLisX1s93FdVSEf3TnHEYHBqy0Qe-8").server("EUW").validated("pending")
-                .validationCode("ASFK732").validationTries(0).buildUser();
-
-        User user3 = new UserBuilder().redditName("Thorasine").summonerName("SecretSmurf")
-                .summonerId("dDsBSlaEk34758KbRuwTnydTNaC1nZQZ5kGOwboGfbb-Zz4").server("NA").validated("validated")
-                .validationCode("TEST754").validationTries(0).buildUser();
-
-        User user4 = new UserBuilder().redditName("Vjostar").summonerName("Vjostar")
-                .summonerId("wWb-WB9G7sMRNTKCjnze7MX84wkIEJhmqe-2JsRwuu_WsEk").server("EUW").validated("pending")
-                .validationCode("QWERTY").validationTries(0).buildUser();
-
-        database.addUser(user1);
-        database.addUser(user2);
-        database.addUser(user3);
-        database.addUser(user4);
+        if (setUpIsDone) {
+            database.dropTable();
+            database.createTable();
+            database.addUser(users.get(0));
+            return;
+        }
+        final List<SpectatorParticipant> participants = new SpectatorBuilder().withPlatform(Platform.EUW1).getFeaturedGames().get(0).getParticipants();
+        for (SpectatorParticipant p : participants) {
+            Summoner summoner = new SummonerBuilder().withPlatform(Platform.EUW1).withName(p.getSummonerName()).get();
+            User user = new UserBuilder().redditName(summoner.getName()).summonerName(summoner.getName())
+                    .summonerId(summoner.getSummonerId()).server("EUW").validated("pending").validationTries(0)
+                    .validationCode("QWERTY").buildUser();
+            users.add(user);
+        }
+        setUpIsDone = true;
+        setUp();
     }
 
     @Test
-    public void update() {
+    public void validationSessionCompleted() {
         verificationUpdater.scheduledUpdate();
+        assertEquals("Validation tries is not 1.", 1, database.getAllUsers().get(0).getValidationTries());
+        assertEquals("Validation tries is not pending.", "pending", database.getAllUsers().get(0).getValidated());
     }
+
+    @Test
+    public void userSetToFailedAfterManyTries() {
+        User user = database.getAllUsers().get(0);
+        user.setValidationTries(10);
+        database.updateUser(user);
+        verificationUpdater.scheduledUpdate();
+        assertEquals("User didn't change from pending to failed after 11 tries.", "failed", database.getAllUsers().get(0).getValidated());
+    }
+
+    @Test
+    public void updatedMultipleUsersSuccessfully(){
+        database.addUser(users.get(1));
+        database.addUser(users.get(2));
+        List<User> dbUsers = database.getAllUsers();
+        User user2 = dbUsers.get(1);
+        user2.setValidationTries(10);
+        database.updateUser(user2);
+        User user3 = dbUsers.get(2);
+        user3.setValidated("validated");
+        int user3Tries = user3.getValidationTries();
+        database.updateUser(user3);
+
+        verificationUpdater.scheduledUpdate();
+        assertEquals("Validation tries is not 1 for user1.", 1, database.getAllUsers().get(0).getValidationTries());
+        assertEquals("Validation tries is not pending. for user1", "pending", database.getAllUsers().get(0).getValidated());
+        assertEquals("Validation tries is not 11 for user2.", 11, database.getUserById(user2.getId()).getValidationTries());
+        assertEquals("Validation tries is not failed for user2.", "failed", database.getUserById(user2.getId()).getValidated());
+        assertEquals("User3 got updated despite being validated.", user3Tries, database.getUserById(user3.getId()).getValidationTries());
+        assertEquals("Validation tries is not validated for user3", "validated", database.getUserById(user3.getId()).getValidated());
+    }
+
 }
